requirements.txt
flask==3.0.3
flask-socketio==5.3.6
pyaudio==0.2.14
faster-whisper==1.0.1
llama-cpp-python==0.2.75
python-dotenv==1.0.1
